Collections: add isCached to CoreData and increase duration from Int16 to Int64

+++ AmperfyKit/Storage/ManagedObjects/ArtistMO+CoreDataProperties.swift
@@ -32,8 +32,8 @@ extension ArtistMO {
     @NSManaged public var albumCount: Int16
     @NSManaged public var remoteAlbumCount: Int16
     @NSManaged public var songCount: Int16
-    @NSManaged public var duration: Int16
-    @NSManaged public var remoteDuration: Int16
+    @NSManaged public var duration: Int64
+    @NSManaged public var remoteDuration: Int64
     @NSManaged public var name: String?
     @NSManaged public var albums: NSOrderedSet?
     @NSManaged public var genre: GenreMO?
